// test/types/DexCore.ts
export type Tokens = {
  token_a: Token;
  token_b: Token;
};

export type SwapSlice = {
  direction: { a_to_b: undefined } | { b_to_a: undefined };
  pair_id: BigNumber;
};

export type LaunchExchange = {
  pair: Tokens;
  token_a_in: BigNumber;
  token_b_in: BigNumber;
  shares_receiver: string;
  candidate: string;
  deadline: string;
};